move cors decorator + code to api app
