+++ corehq/apps/api/cors.py
@@ -1,0 +1,5 @@
+
+
+ACCESS_CONTROL_ALLOW = 'Allow'
+ACCESS_CONTROL_ALLOW_HEADERS = 'Access-Control-Allow-Headers'
+ACCESS_CONTROL_ALLOW_ORIGIN = 'Access-Control-Allow-Origin'
