+++ corehq/apps/api/decorators.py
@@ -1,7 +1,9 @@
 import base64
+from functools import wraps
 
 from django.http import HttpResponse
 
+from corehq.apps.api.cors import ACCESS_CONTROL_ALLOW_ORIGIN, ACCESS_CONTROL_ALLOW_HEADERS, ACCESS_CONTROL_ALLOW
 from corehq.apps.api.models import ApiUser
 
 
@@ -21,3 +23,26 @@ def api_user_basic_auth(permission, realm=''):
             return response
         return wrapper
     return real_decorator
+
+
+def allow_cors(allowed_methods):
+    allowed_methods = allowed_methods or []
+    # always allow options
+    allowed_methods = allowed_methods + ['OPTIONS']
+
+    def decorator(view_func):
+        @wraps(view_func)
+        def wrapped_view(request, *args, **kwargs):
+            if request.method == "OPTIONS":
+                response = HttpResponse()
+                response[ACCESS_CONTROL_ALLOW_ORIGIN] = '*'
+                response[ACCESS_CONTROL_ALLOW_HEADERS] = 'Content-Type, Authorization'
+                response[ACCESS_CONTROL_ALLOW] = ', '.join(allowed_methods)
+                return response
+            response = view_func(request, *args, **kwargs)
+            if request.method in allowed_methods:
+                response[ACCESS_CONTROL_ALLOW_ORIGIN] = '*'
+                response[ACCESS_CONTROL_ALLOW_HEADERS] = 'Content-Type, Authorization'
+            return response
+        return wrapped_view
+    return decorator
