+++ corehq/apps/domain/decorators.py
@@ -692,31 +692,3 @@ def track_domain_request(calculated_prop):
         return _wrapped
 
     return _dec
-
-
-ACCESS_CONTROL_ALLOW = 'Allow'
-ACCESS_CONTROL_ALLOW_HEADERS = 'Access-Control-Allow-Headers'
-ACCESS_CONTROL_ALLOW_ORIGIN = 'Access-Control-Allow-Origin'
-
-
-def allow_cors(allowed_methods):
-    allowed_methods = allowed_methods or []
-    # always allow options
-    allowed_methods = allowed_methods + ['OPTIONS']
-
-    def decorator(view_func):
-        @wraps(view_func)
-        def wrapped_view(request, *args, **kwargs):
-            if request.method == "OPTIONS":
-                response = HttpResponse()
-                response[ACCESS_CONTROL_ALLOW_ORIGIN] = '*'
-                response[ACCESS_CONTROL_ALLOW_HEADERS] = 'Content-Type, Authorization'
-                response[ACCESS_CONTROL_ALLOW] = ', '.join(allowed_methods)
-                return response
-            response = view_func(request, *args, **kwargs)
-            if request.method in allowed_methods:
-                response[ACCESS_CONTROL_ALLOW_ORIGIN] = '*'
-                response[ACCESS_CONTROL_ALLOW_HEADERS] = 'Content-Type, Authorization'
-            return response
-        return wrapped_view
-    return decorator
